2019-06-19 | Comment

+++ django/myform/boards/views.py
@@ -1,6 +1,10 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Board
-from .forms import BoardForm
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST
+
+from .models import Board, Comment
+from .forms import BoardForm, CommentForm
+
 from IPython import embed
 
 
@@ -9,6 +13,8 @@ def index(request):
     context = {'boards':boards}
     return render(request, 'boards/index.html', context)
 
+
+@login_required
 def create(request):
     if request.method == "POST":
         form = BoardForm(request.POST)
@@ -16,7 +22,9 @@ def create(request):
         ## 유효성 검사 단계: Null값이나 type 오류를 찾아냄
         ## is_valid(): True / False 형태로 결과를 나타냄
         if form.is_valid():
-            board = form.save()
+            board = form.save(commit=False)
+            board.user = request.user
+            board.save()
             ## cleaned_data: True / False를 정제해서 나타냄
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
@@ -32,39 +40,67 @@ def create(request):
 def detail(request, board_pk):
     # board = Board.objects.get(pk=board_pk)
     board = get_object_or_404(Board, pk=board_pk)
-    context = {'board':board}
+    comments = board.comment_set.all()
+    comment_form = CommentForm()
+    context = {'board':board,
+               'comments':comments,
+               'comment_form':comment_form,}
     return render(request, 'boards/detail.html', context)
 
 
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == "POST":
-        board.delete()
-        return redirect('boards:index')
-
-    context = {'board':board}
-    return render(request, 'boards/detail.html', context)
-    # else:
-    #     context = {'board':board}
-    #     return render(request, 'boards/detail.html', context)
-
-
-def update(request, board_pk):
-    board = get_object_or_404(Board, pk=board_pk)
-    if request.method == "POST":
-        form = BoardForm(request.POST, instance=board)
-        if form.is_valid():
-            form.save()
-            # board.title = form.cleaned_data.get('title')
-            # board.content = form.cleaned_data.get('content')
-            # board.save()
+    if board.user == request.user:
+        if request.method == "POST":
+            board.delete()
+            return redirect('boards:index')
+        else:
             return redirect('boards:detail', board.pk)
     else:
-        # initial=board.__dict__: 기존에 사용자가 작성한 text를 불러오는 함수
-        form = BoardForm(instance=board)
+        return redirect('boards:index')
+
+
+@login_required
+def edit(request, board_pk):
+    board = get_object_or_404(Board, pk=board_pk)
+    if board.user == request.user:
+        if request.method == "POST":
+            form = BoardForm(request.POST, instance=board)
+            if form.is_valid():
+                form.save()
+                # board.title = form.cleaned_data.get('title')
+                # board.content = form.cleaned_data.get('content')
+                # board.save()
+                return redirect('boards:detail', board.pk)
+        else:
+            # initial=board.__dict__: 기존에 사용자가 작성한 text를 불러오는 함수
+            form = BoardForm(instance=board)
+    else:
+        return redirect('boards:index')
     context = {'form': form, 'board':board}
     return render(request, 'boards/form.html', context)
 
+
+@login_required
+@require_POST
+def comments_create(request, board_pk):
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False)
+        comment.user = request.user
+        comment.board_id = board_pk
+        comment.save()
+    return redirect('boards:detail', board_pk)
+
+
+@login_required
+@require_POST
+def comments_delete(request, board_pk, comment_pk):
+    comment = get_object_or_404(Comment, pk=comment_pk)
+    if request.user != comment.user:
+        return redirect('boards:detail', board_pk)
+    comment.delete()
+    return redirect('boards:detail', board_pk)
 
 
 def photo(request):
